Implement SafetyRace plugin
This plugin lets users reshuffle their decks when reloading specific
events. It accomplishes fairly straightforwardly, by watching for the
game loading and then manually reshuffling the player and npc decks.

Additionally it modifies the menu buttons to make it obvious that a mod
is loaded: "Begin Story" becomes "Begin Race" "Continue Story" becomes
"Continue Story*" and, if one of the matched events is being reloaded,
"Continue Story" becomes "Reshuffle"

There's some fiddliness with the transparency on the menu button text,
caused, I think, by the fade-in animation on the buttons when the game
first starts. The end result is that all the TextMeshPro text objects
have an alpha of 0 (transparent), but this doesn't get applied until we
change the literal text. So we also need to make sure to set the alpha
to 1 (except not during that initial fade in animation since that looks
wonky)