Add "Retry with current rng" feature
This change moves the "Reshuffle" menu button off to the side, and adds
a "Retry Conversation" menu button in the old location.
The "Retry Conversation" button allows the player to re-do the
conversation they just quit-to-menu from with the same RNG seed (and
thus the same card order) they were just playing with.

This change alters the way the plugin works under the hood:
instead of reaching into the game state and manually calling
`ShuffleInPlace` on the player and npc decks, it instead advances the
RNG state before the game's regular deck shuffling happens, saving that
new state, each time the player hits "Reshuffle"

Co-authored-by: rose.as.romeo