Move Reshuffle button inline with normal menu buttons
This change moves the menu buttons around so that they do not get cut
off, above or below, when we add an additional button on top.
Additionally, the custom "Reshuffle" button is connected to the
text-schooching animations so that it does not look out of place

This change also adjusts the highlight underscore that appears on
mouseover to match the width of our custom text.

Co-Authored-By: rose.as.romeo